requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
backcall==0.2.0
cachetools==4.2.4
certifi==2025.1.31
charset-normalizer==2.0.12
dataclasses==0.8
decorator==5.1.1
entrypoints==0.4
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
future==1.0.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
huggingface-hub==0.4.0
idna==3.10
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
joblib==1.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
Markdown==3.3.7
multidict==5.2.0
nest-asyncio==1.6.0
numpy==1.19.5
oauthlib==3.2.2
packaging==21.3
pandas==1.1.5
parso==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
pydicom==2.3.1
Pygments==2.14.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.10.post0
pytz==2025.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.24.2
scipy==1.5.4
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
timm==0.6.12
torch==1.10.1
torchaudio==0.10.1
torchmetrics==0.8.2
torchvision==0.11.2
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0
